fix: fix date picker with timezone
